requirements.txt
fastapi
uvicorn
python-dotenv
fpdf
PyPDF2